// src/dorc-web/src/apis/dorc-api/models/ScriptAuditApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface ScriptAuditApiModel
 */
export interface ScriptAuditApiModel {
    /**
     * @type {number}
     * @memberof ScriptAuditApiModel
     */
    Id?: number;
    /**
     * @type {number}
     * @memberof ScriptAuditApiModel
     */
    ScriptId?: number | null;
    /**
     * @type {string}
     * @memberof ScriptAuditApiModel
     */
    ScriptName?: string | null;
    /**
     * @type {string}
     * @memberof ScriptAuditApiModel
     */
    FromValue?: string | null;
    /**
     * @type {string}
     * @memberof ScriptAuditApiModel
     */
    ToValue?: string | null;
    /**
     * @type {string}
     * @memberof ScriptAuditApiModel
     */
    UpdatedBy?: string | null;
    /**
     * @type {string}
     * @memberof ScriptAuditApiModel
     */
    UpdatedDate?: string;
    /**
     * @type {string}
     * @memberof ScriptAuditApiModel
     */
    Type?: string | null;
    /**
     * @type {string}
     * @memberof ScriptAuditApiModel
     */
    ProjectNames?: string | null;
}