// src/dorc-web/src/apis/dorc-api/models/GetScriptsAuditListResponseDto.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type {
    ScriptAuditApiModel,
} from './';

/**
 * @export
 * @interface GetScriptsAuditListResponseDto
 */
export interface GetScriptsAuditListResponseDto {
    /**
     * @type {number}
     * @memberof GetScriptsAuditListResponseDto
     */
    CurrentPage?: number;
    /**
     * @type {number}
     * @memberof GetScriptsAuditListResponseDto
     */
    TotalItems?: number;
    /**
     * @type {number}
     * @memberof GetScriptsAuditListResponseDto
     */
    TotalPages?: number;
    /**
     * @type {Array<ScriptAuditApiModel>}
     * @memberof GetScriptsAuditListResponseDto
     */
    Items?: Array<ScriptAuditApiModel> | null;
}