// src/dorc-web/src/apis/dorc-api/apis/ScriptsAuditApi.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { Observable } from 'rxjs';
import type { AjaxResponse } from 'rxjs/ajax';
import { BaseAPI } from '../runtime';
import type { OperationOpts, HttpHeaders, HttpQuery } from '../runtime';
import type {
    GetScriptsAuditListResponseDto,
    PagedDataOperators,
} from '../models';

export interface ScriptsAuditPutRequest {
    page?: number;
    limit?: number;
    useAndLogic?: boolean;
    pagedDataOperators?: PagedDataOperators;
}

/**
 * no description
 */
export class ScriptsAuditApi extends BaseAPI {

    /**
     */
    scriptsAuditPut({ page, limit, useAndLogic, pagedDataOperators }: ScriptsAuditPutRequest): Observable<GetScriptsAuditListResponseDto>
    scriptsAuditPut({ page, limit, useAndLogic, pagedDataOperators }: ScriptsAuditPutRequest, opts?: OperationOpts): Observable<AjaxResponse<GetScriptsAuditListResponseDto>>
    scriptsAuditPut({ page, limit, useAndLogic, pagedDataOperators }: ScriptsAuditPutRequest, opts?: OperationOpts): Observable<GetScriptsAuditListResponseDto | AjaxResponse<GetScriptsAuditListResponseDto>> {

        const headers: HttpHeaders = {
            'Content-Type': 'application/json',
            // oauth required
            ...(this.configuration.accessToken != null
                ? { Authorization: typeof this.configuration.accessToken === 'function'
                    ? this.configuration.accessToken('oauth2', ['dorc-api-np.manage'])
                    : this.configuration.accessToken }
                : undefined
            ),
        };

        const query: HttpQuery = {};

        if (page != null) { query['page'] = page; }
        if (limit != null) { query['limit'] = limit; }
        if (useAndLogic != null) { query['useAndLogic'] = useAndLogic; }

        return this.request<GetScriptsAuditListResponseDto>({
            url: '/ScriptsAudit',
            method: 'PUT',
            headers,
            query,
            body: pagedDataOperators,
        }, opts?.responseOpts);
    };

}